tests/e2e-quality/pyproject.toml
[project]
name = "e2e-quality"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pytest>=8.0,<9",
    "httpx>=0.27,<1",
    "pyyaml>=6.0,<7",
    "pytest-html>=4.0,<5",
]

[tool.pytest.ini_options]
testpaths = ["."]
addopts = "-ra"
